+++ src/main/java/domain/card/CardRank.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum CardRank {
     ACE(1, "A"),
