+++ src/main/java/domain/game/GameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public enum GameResult {
     WIN,
