+++ src/test/java/domain/DealerTest.java
@@ -2,6 +2,9 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.Deck;
+import domain.participants.Dealer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
