+++ src/test/java/domain/PlayerNameTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.participants.PlayerName;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
